moved ios & macos setup to separate files
